play

+++ utility/utils/xplay/xplay.go
@@ -7,7 +7,7 @@ import (
 )
 
 // 开奖结果转换成比分总数
-func OpenResutToTotal(openResult string) (int, error) {
+func OpenResutToTotal(openResult string) (float64, error) {
 	two, i, err := OpenResutToTwo(openResult)
 	if err != nil {
 		return 0, err
@@ -16,10 +16,10 @@ func OpenResutToTotal(openResult string) (int, error) {
 }
 
 // 开奖结果转换成两个比分
-func OpenResutToTwo(openResult string) (int, int, error) {
+func OpenResutToTwo(openResult string) (float64, float64, error) {
 	split := strings.Split(openResult, "-")
 	if len(split) < 2 {
 		return 0, 0, fmt.Errorf("长度错误")
 	}
-	return gconv.Int(split[0]), gconv.Int(split[1]), nil
+	return gconv.Float64(split[0]), gconv.Float64(split[1]), nil
 }
